requirements.txt
# Core Deep Learning Libraries
torch==2.2.0
torchvision==0.17.0
torchaudio==2.2.0  # Optional

# General Dependencies
numpy==1.26.4
scikit-learn==1.4.2
matplotlib==3.8.4
seaborn==0.13.2
pandas==2.2.2
tqdm==4.66.2
pillow==10.2.0
opencv-python==4.9.0.80
scipy==1.13.0

opencv-contrib-python==4.9.0.80  # Required for DualTVL1 optical flow

# PyTorch I3D
pytorch-i3d  # Local module, install via pip install -e . or include manually

# Utility Libraries
tqdm==4.66.2

# Others (often auto-installed but pinned here for reproducibility)
joblib==1.4.2
typing-extensions==4.11.0
packaging==24.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
six==1.16.0
jinja2==3.1.3
fsspec==2024.3.1
filelock==3.13.1
sympy==1.12
mpmath==1.3.0
tzdata==2024.1

# Visualization extras
cycler==0.12.1
contourpy==1.2.1
fonttools==4.50.0
kiwisolver==1.4.5


plotly==6.0.1